refactor(throttle): move inline throttler to utils; revert package-lock.json to base; tests passing
